# docs/swagger.yaml
basePath: /fab
definitions:
...
  gintool.ApiResponse:
    properties:
      code:
        description: 状态码
        type: integer
      err_msg:
        description: 内部错误详情
        type: string
      message:
        description: 状态短语
        type: string
      result:
        description: 数据结果集
        type: object
    type: object
  ledger.queryBlockByHashRequest:
    properties:
      channelid:
        example: mychannel
        type: string
      hash:
        example: 0xacedababacedababacedababacedababacedababacedababacedababacedabab
        type: string
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
    required:
    - channelid
    - hash
    - orgmsps
    - orgnames
    type: object
  ledger.queryBlockByTxIdRequest:
    properties:
      channelid:
        example: mychannel
        type: string
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
      txid:
        example: 0xacedababacedababacedababacedababacedababacedababacedababacedabab
        type: string
    required:
    - channelid
    - orgmsps
    - orgnames
    - txid
    type: object
  ledger.queryBlockNumRequest:
    properties:
      channelid:
        example: mychannel
        type: string
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
    required:
    - orgmsps
    - orgnames
    type: object
  ledger.queryBlockRequest:
    properties:
      blocknum:
        example: 0
        type: integer
      channelid:
        example: mychannel
        type: string
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
    required:
    - channelid
    - orgmsps
    - orgnames
    type: object
  ledger.queryConfigRequest:
    properties:
      channelid:
        example: mychannel
        type: string
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
    required:
    - channelid
    - orgmsps
    - orgnames
    type: object
  ledger.queryLedgerRequest:
    properties:
      channelid:
        example: mychannel
        type: string
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
    required:
    - channelid
    - orgmsps
    - orgnames
    type: object
  ledger.queryTxRequest:
    properties:
      channelid:
        example: mychannel
        type: string
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
      txid:
        example: 0xacedababacedababacedababacedababacedababacedababacedababacedabab
        type: string
    required:
    - orgmsps
    - orgnames
    - txid
    type: object
...
host: localhost:6922
info:
  contact: {}
  description: 描述:fabapi 接口文档.
  license: {}
  title: fabapi API文档
  version: 1.1.0
paths:
...
  /ledger/queryblock:
    post:
      consumes:
      - application/json
      description: 功能：查询块管理
      parameters:
      - description: 查询块管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/ledger.queryBlockRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 查询块管理
      tags:
      - 查询块管理
  /ledger/queryblockbyhash:
    post:
      consumes:
      - application/json
      description: 功能：通过哈希查询块管理
      parameters:
      - description: 通过哈希查询块管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/ledger.queryBlockByHashRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 通过哈希查询块管理
      tags:
      - 通过哈希查询块管理
  /ledger/queryblockbytxid:
    post:
      consumes:
      - application/json
      description: 功能：通过交易id查询块管理
      parameters:
      - description: 通过交易id查询块管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/ledger.queryBlockByTxIdRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 通过交易id查询块管理
      tags:
      - 通过交易id查询块管理
  /ledger/queryblocknum:
    post:
      consumes:
      - application/json
      description: 功能：查询块高管理
      parameters:
      - description: 查询块高管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/ledger.queryBlockNumRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 查询块高管理
      tags:
      - 查询块高
  /ledger/queryconfig:
    post:
      consumes:
      - application/json
      description: 功能：查询配置块管理
      parameters:
      - description: 查询配置块管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/ledger.queryConfigRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 查询配置块管理
      tags:
      - 查询配置块管理
  /ledger/queryledger:
    post:
      consumes:
      - application/json
      description: 功能：查询账本管理
      parameters:
      - description: 查询账本管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/ledger.queryLedgerRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 查询账本管理
      tags:
      - 查询账本管理
  /ledger/querytx:
    post:
      consumes:
      - application/json
      description: 功能：查询交易管理
      parameters:
      - description: 查询交易管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/ledger.queryTxRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 查询交易管理
      tags:
      - 查询交易管理
...
swagger: "2.0"